# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Loopy Bot Admin Panel API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface OkResponse {
  ok: boolean;
}

export interface AdminUser {
  id: string;
  username: string;
  /** @nullable */
  avatarUrl: string | null;
}

export interface BotStats {
  guildCount: number;
  totalMembers: number;
  commandCount: number;
  uptimeSeconds: number;
  botUsername: string;
  /** @nullable */
  botAvatarUrl: string | null;
}

export interface BotGuild {
  id: string;
  name: string;
  /** @nullable */
  iconUrl: string | null;
  memberCount: number;
  ownerId: string;
  /** @nullable */
  joinedAt: string | null;
}

export interface GuildMember {
  userId: string;
  username: string;
  displayName: string;
  /** @nullable */
  avatarUrl: string | null;
  isBot: boolean;
  /** @nullable */
  joinedAt: string | null;
  roles: string[];
}

export interface KickMemberBody {
  reason?: string;
}

export interface GuildChannel {
  id: string;
  name: string;
  type: string;
  position: number;
  /** @nullable */
  parentId: string | null;
  /** @nullable */
  parentName: string | null;
  /** @nullable */
  memberCount: number | null;
}

export interface BotCommand {
  id: string;
  name: string;
  description: string;
  type: number;
  /** @nullable */
  guildId: string | null;
}

export interface InviteTarget {
  id: number;
  guildId: string;
  /** @nullable */
  label: string | null;
  botAlreadyIn: boolean;
  inviteUrl: string;
  addedAt: string;
}

export interface InviteTargetInput {
  /** @minLength 1 */
  guildId: string;
  label?: string;
}

export interface SendMessageBody {
  channelId: string;
  content: string;
}

export interface SentMessage {
  ok: boolean;
  messageId: string;
}

export interface InviteUrl {
  url: string;
  guildId: string;
}

export type GetGuildMembersParams = {
limit?: number;
};